Fix CI/CD: Add size_t casts in Labellist setN and reserve methods
Fixed sign conversion warnings in lparser.h:
- setN(): Cast new_n to size_t before passing to vec.resize()
- reserve(): Cast capacity to size_t before passing to vec.reserve()

Both methods take int parameters but the underlying std::vector
methods expect size_t (unsigned long), causing sign conversion warnings.

+++ src/compiler/lparser.h
@@ -226,11 +226,11 @@ public:
   inline int getSize() const noexcept { return static_cast<int>(vec.capacity()); }
 
   /* Modifying size */
-  inline void setN(int new_n) { vec.resize(new_n); }
+  inline void setN(int new_n) { vec.resize(static_cast<size_t>(new_n)); }
 
   /* Direct vector access for modern operations */
   inline void push_back(const Labeldesc& desc) { vec.push_back(desc); }
-  inline void reserve(int capacity) { vec.reserve(capacity); }
+  inline void reserve(int capacity) { vec.reserve(static_cast<size_t>(capacity)); }
   inline Labeldesc& operator[](int index) { return vec[index]; }
   inline const Labeldesc& operator[](int index) const { return vec[index]; }
